# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: DocFlow API
  version: 1.0.0
  description: |
    DocFlow 文档管理服务 API。

    ## 认证
    - 业务端点要求 `Authorization: Bearer <access_token>`（HS256 JWT，subject 为用户 UUID）。
    - 登录/刷新通过 HttpOnly Cookie `refresh_token`（Path=/api/v1/auth/refresh，SameSite=Lax）轮换会话。

    ## 错误与限流
    - 错误响应统一为 `{error, code?}`；`code` 为机器可读错误码（如 NAME_CONFLICT、QUARANTINED）。
    - 认证接口按 user_id+IP 每分钟限流；登录按 IP+邮箱前缀哈希更严格限流；
      公开分享接口按 IP 独立限流；超限返回 429 并带 Retry-After。
    - 文件元数据响应带 ETag（基于 updated_at）。
servers:
  - url: /api/v1
    description: API 基础路径（契约测试据此归一化路由）
security:
  - bearerAuth: []
...
paths:
...
  /uploads:
    post:
      tags: [Uploads]
      operationId: createUpload
      summary: 创建上传会话（自定义分片 API）
      description: >-
        与 tus 端点并存的自定义 API。注意：此处服务启动失败类错误（如超出大小上限）
        一律映射 400，而非 tus 端点的 413。
        携带 file_id 时为「覆盖为新版本」会话：目标须为 type=file 且当前用户有
        CanWrite 权限（个人 owner / 团队 editor+），name 与 parent_id 被忽略，
        沿用目标文件现有名称与父目录；Complete 成功后内容成为该文件的新版本，
        并按 MAX_VERSIONS_PER_FILE（默认 5，upload.max_versions_per_file 热调）
        裁剪历史版本（upload.version_retention_days 时间窗内的版本不因数量
        裁剪删除；0=不启用）。
        门控：扩展名命中 upload.blocked_extensions 黑名单时 400；每用户活跃
        （非终态）会话达到 upload.max_concurrent_uploads_per_user 上限时 429
        （DB COUNT 计数，多实例共享）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, size]
              properties:
                name: { type: string }
                size: { type: integer, format: int64, minimum: 0 }
                expected_sha256:
                  { type: string, description: 期望内容 SHA-256（64 位十六进制，大小写不敏感） }
                parent_id:
                  { type: string, format: uuid, description: 目标目录；缺省为用户根目录 }
                file_id:
                  { type: string, format: uuid, description: 覆盖目标文件 ID；指定时 name/parent_id 被忽略 }
      responses:
        '201':
          description: 会话已创建
          headers:
            Location:
              description: '/api/v1/uploads/{id}'
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSession' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: >-
            无 CanWrite 权限（团队 viewer/非成员）、版本覆盖能力未启用，或
            存储配额超限——后者 body 为 {error, code: QUOTA_EXCEEDED}
            （C3：已用含回收站软删文件，彻底删除后才释放配额）。
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: file_id 目标不存在（或个人文件非 owner，不泄露存在性）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /uploads/{id}:
    get:
      tags: [Uploads]
      operationId: getUpload
      summary: 查询上传会话
      description: 轮询上传状态（含后台自动完成后的终态）；响应为 UploadSession 的子集（不含 retry_count）。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 会话状态
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSession' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      tags: [Uploads]
      operationId: patchUpload
      summary: 追加上传数据
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Upload-Offset
          in: header
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
          description: 必须等于服务端当前 offset，否则 409
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: 追加成功；响应头 Upload-Offset 为新偏移
          headers:
            Upload-Offset: { schema: { type: integer, format: int64 } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSession' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /uploads/{id}/complete:
    post:
      tags: [Uploads]
      operationId: completeUpload
      summary: 完成上传
      description: >-
        校验哈希 → 病毒扫描（如启用）→ 落库为文件版本。终态幂等：available 重复调用
        返回 200；quarantined / failed 返回 409（code 分别为 QUARANTINED / FAILED）。
        扩展名黑名单复检（会话期间 upload.blocked_extensions 可能变更）：命中时
        400 且会话置 failed。
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 完成后的会话（status=available）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadSession' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '扫描被隔离 / 校验失败 / 会话不可完成；code: QUARANTINED 或 FAILED'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Authorization: Bearer <token>，两类凭证等价可用：
        ① HS256 JWT（subject 为用户 UUID，过期时间以服务配置为准）；
        ② PAT 个人访问令牌：dfpat_ + 43 字符 URL-safe base64（经
        POST /tokens 创建，撤销/过期即失效；无 refresh 语义）。
  responses:
    BadRequest:
      description: 请求参数或请求体非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 缺失或无效的访问令牌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: 资源不存在或不属于当前用户（不泄露存在性）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 状态冲突（名称冲突、偏移不匹配、文件无当前版本等）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: 触发限流；响应头 Retry-After 为当前窗口剩余秒数
      headers:
        Retry-After:
          description: 建议等待秒数
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 人类可读错误信息
        code:
          type: string
          description: >-
            机器可读错误码；当前实现使用 NAME_CONFLICT / NOT_DELETED / PARENT_DELETED /
            QUARANTINED / FAILED / TOTP_REQUIRED（POST /auth/login 对已启用两步
            验证账号的 401，error=totp_required）/ ACCOUNT_LOCKED（登录与
            refresh 遇锁定期账号的 423）/ QUOTA_EXCEEDED（上传建会话配额
            超限的 403）/ CSRF_REJECTED（refresh/logout 同源校验失败的 403）
        status:
          type: string
          description: 上传相关错误附带的会话状态（仅 uploads complete 端点）
...
    UploadSession:
      type: object
      description: >-
        上传会话；POST /uploads、PATCH /uploads/{id}、POST /uploads/{id}/complete
        返回全量字段，GET /uploads/{id} 省略 retry_count。响应不含
        user_id、expected_sha256、metadata、storage_key（敏感/内部字段）。
      properties:
        id: { type: string, format: uuid }
        parent_id: { type: string, format: uuid }
        name: { type: string }
        size: { type: integer, format: int64, description: 声明的总字节数 }
        offset: { type: integer, format: int64, description: 已接收字节数 }
        status:
          type: string
          enum: [uploading, verifying, scanning, available, quarantined, failed]
        file_id:
          type: [string, 'null']
          format: uuid
          description: >-
            覆盖为新版本会话的目标文件 ID（仅 file_id 创建的会话返回，含 Complete
            完成后的响应）；常规上传会话缺省该字段
        expires_at: { type: string, format: date-time }
        retry_count:
          { type: integer, description: '处理重试次数（GET /uploads/{id} 不返回）' }
        created_at: { type: string, format: date-time }
        completed_at:
          type: [string, 'null']
          format: date-time
          description: 完成时间；未完成为 null 或缺省
...